requirements.txt
# core
requests>=2.28
beautifulsoup4>=4.12
pandas>=2.0
sqlalchemy>=1.4
python-dotenv>=1.0

# scrapers & social
instaloader>=4.9.6
facebook_scraper>=0.2.58
snscrape>=0.4.4

# LLM (local)
transformers>=4.40
torch>=2.0.0         # CPU or GPU; automatic selection
sentence-transformers>=2.2.2

# utility
tqdm>=4.65
uvicorn>=0.20
fastapi>=0.95        # optional if you want to expose endpoints locally
streamlit>=1.26      # optional frontend

# vector db (optional)
faiss-cpu>=1.7.4